% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer.R
\name{boost_lmer}
\alias{boost_lmer}
\title{Boost an lmer Model-Object}
\usage{
boost_lmer(model, tol = 1e-08)
}
\arguments{
\item{model}{and lmer model-object -- the result of a call to \code{lme4::lmer()}}

\item{tol}{tolerance for determining of eigenvalues are negative, zero or
positive}
}
\value{
a list with components
\item{parlist}{list of parameter estimates including beta,
theta, sigma (vectors) and vcov(beta) (matrix)}
\item{A}{the asymptotic covariance matrix of the variance parameters
(theta, sigma)}
\item{Jac_list}{list of Jacobian matrices; gradients of vcov(beta) with
respect to the variance parameters }
}
\description{
Boosting an lme4::lmer model-object involves computing the covariance
matrix of the variance parameters and the gradient (Jacobian) of cov(beta)
with respect to the variance parameters.
}
\examples{
m <- lme4::lmer(Reaction ~ Days + (Days | Subject), sleepstudy)
bm <- lmerTestR:::boost_lmer(m)
names(bm)

}
\author{
Rune Haubo B. Christensen
}
\keyword{internal}
