% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast_utils.R
\name{containment}
\alias{containment}
\title{Determine the Containment Structure for All Terms in a Model}
\usage{
containment(object)
}
\arguments{
\item{object}{a model object, e.g. of class \code{lm} or \code{merMod}.}
}
\value{
a list with one element for each term in the model. Each element/term
is a character vector of terms that the term is contained in.
}
\description{
See \code{\link{term_contain}} for details about containment.
}
\keyword{internal}
