% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contest.R
\name{contest1D}
\alias{contest1D}
\title{Contrast Tests in 1D}
\usage{
contest1D(L, model, ddf = c("Satterthwaite", "KR"))
}
\arguments{
\item{L}{a numeric (contrast) vector of the same length as
\code{fixef(model)}.}

\item{model}{a model object fitted with \code{lmer} from package
\pkg{lmerTestR}, i.e., an object of class \code{\link{lmerModLmerTest}}.}

\item{ddf}{the method for computing the denominator degrees of freedom.
\code{ddf="KR"} uses Kenward-Roger's method.}
}
\value{
A \code{data.frame} with one row and columns with \code{"Estimate"},
\code{"Std. Error"}, \code{"t value"}, \code{"df"}, and \code{"Pr(>|t|)"}
(p-value)
}
\description{
Compute the test of a one-dimensional (vector) contrast in a
linear mixed model fitted with lmer from package \pkg{lmerTestR}.
The contrast should specify a linear function of the
mean-value parameters, beta. Satterthwaite's method is used to compute the
denominator df for the t-test.
}
\examples{

# Fit model using lmer with data from the lme4-package:
fm <- lmer(Reaction ~ Days + (1 + Days|Subject), sleepstudy)
# Note that summary do not contain any tests/p-values:
coef(summary(fm))
# We can get these tests with:
contest1D(c(1, 0), fm) # Test for Intercept
contest1D(c(0, 1), fm) # Test for Days

}
\seealso{
\code{\link{contestMD}} for multi degree-of-freedom contrast tests.
}
\author{
Rune Haubo B. Christensen
}
