% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contest.R
\name{contestMD}
\alias{contestMD}
\title{Multiple Degrees-of-Freedom Contrast Tests}
\usage{
contestMD(L, model, ddf = c("Satterthwaite", "KR"),
  eps = sqrt(.Machine$double.eps))
}
\arguments{
\item{L}{a contrast matrix with nrow >= 1 and ncol ==
\code{length(fixef(model))}.}

\item{model}{a model object fitted with \code{lmer} from package
\pkg{lmerTestR}, i.e., an object of class \code{\link{lmerModLmerTest}}.}

\item{ddf}{the method for computing the denominator degrees of freedom and
F-statistics. \code{ddf="KR"} uses Kenward-Roger's method.}

\item{eps}{tolerance on eigenvalues to determine if an eigenvalue is
positive. The number of positive eigenvalues determine the rank of
L and the numerator df of the F-test.}
}
\value{
a \code{data.frame} with one row and columns with \code{"Sum Sq"},
\code{"Mean Sq"}, \code{"F value"}, \code{"NumDF"} (numerator df),
\code{"DenDF"} (denominator df) and \code{"Pr(>F)"} (p-value).
}
\description{
Compute the multi degrees-of-freedom test in a linear mixed model fitted
by \code{\link{lmer}}. The contrast (L) specifies a linear function of the
mean-value parameters, beta. Satterthwaite's method is used to compute the
denominator df for the F-test.
}
\details{
Note: NumDF = row-rank(L) is determined automatically so row rank-deficient L
are allowed. One-dimensional contrasts are also allowed (L has 1 row).
}
\examples{

fm <- lmer(Reaction ~ Days + I(Days^2) + (1|Subject) + (0+Days|Subject),
           sleepstudy)
# Define 2-df contrast - since L has 2 (linearly independent) rows
# the F-test is on 2 (numerator) df:
L <- rbind(c(0, 1, 0), # Note: ncol(L) == length(fixef(fm))
           c(0, 0, 1))
# Make the 2-df F-test of any effect of Days:
contestMD(L, fm)
# Make the 1-df F-test of the effect of Days^2:
contestMD(L[2, , drop=FALSE], fm)
# Same test, but now as a t-test instead:
contest1D(L[2, , drop=TRUE], fm)

}
\seealso{
\code{\link{contest1D}} for tests of 1-dimensional contrasts.
}
\author{
Rune Haubo B. Christensen
}
