% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast_utils.R
\name{contrast_type3SAS}
\alias{contrast_type3SAS}
\title{Compute Type III Contrast Matrix for a Model Term}
\usage{
contrast_type3SAS(term, terms, data_classes, L, eps = 1e-08)
}
\arguments{
\item{term}{the name of a model term; one of
\code{attr(terms(model), "term.labels")}.}

\item{terms}{a terms object (using \code{fixed.only=TRUE}).}

\item{data_classes}{\code{attr(terms(model, fixed.only=FALSE), "dataClasses")}.}

\item{L}{the pp x pp matrix with pp-p zero-rows indicating the relation between the
pp parameters and p coefficients.}

\item{eps}{tolererance for non-zero pivots}
}
\value{
a pp-vector or qi x pp matrix with the contrast for the term.
}
\description{
Experimental - not extensively tested.
}
\details{
Compute the type III qi x pp contrast matrix for 'term' where qi is the
numerater df for 'term' and pp is the number of parameters in the
over-parameterized model notation (and also equals the number of columns in
the 'full' rank-deficient design matrix). if qi=1 the result is a qi-vector.
}
\author{
Rune Haubo B. Christensen based on
\code{lmerTest:::makeContrastType3SAS}
}
\keyword{internal}
