% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast_utils.R
\name{doolittle}
\alias{doolittle}
\title{Doolittle Decomposition}
\usage{
doolittle(x, eps = 1e-06)
}
\arguments{
\item{x}{a numeric square matrix with at least 2 columns/rows.}

\item{eps}{numerical tolerance on the whether to normalize with components
in \code{L} with the diagonal elements of \code{U}.}
}
\value{
a list with two matrices of the same dimension as \code{x}:
\item{L}{lower-left unit-triangular matrix}
\item{U}{upper-right triangular matrix (\emph{not} unit-triangular)}
}
\description{
Doolittle Decomposition
}
\keyword{internal}
