% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast_utils.R
\name{general_L}
\alias{general_L}
\title{Compute a 'General Set of Estimable Functions'}
\usage{
general_L(rdX, is_coef)
}
\arguments{
\item{rdX}{a 'full' rank-deficient design matrix.}

\item{is_coef}{an indicator vector of the same length as \code{fixef(model)}
which indicates which columns in \code{rdX} correspond \code{fixef(model)}.}
}
\value{
The L-matrix
}
\description{
Computes the so-called L-matrix (in SAS terminology).
Experimental - not extensively tested.
}
\author{
Rune Haubo B. Christensen
}
\keyword{internal}
