% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type12.R
\name{get_contrasts_type1}
\alias{get_contrasts_type1}
\title{Type I ANOVA table contrasts}
\usage{
get_contrasts_type1(model, keep_intercept = FALSE)
}
\arguments{
\item{model}{a model object with \code{terms} and \code{model.matrix} methods.}

\item{keep_intercept}{defaults to \code{FALSE}. If \code{TRUE} a contrast
for the intercept is included in the returned list of contrast matrices.}
}
\value{
List of contrast matrices - one contrast matrix for each model term.
}
\description{
Type I ANOVA table contrasts
}
\author{
Rune Haubo B. Christensen
}
\keyword{internal}
