% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type3.R
\name{get_contrasts_type3}
\alias{get_contrasts_type3}
\title{Compute Type III Contrast Matrices}
\usage{
get_contrasts_type3(model)
}
\arguments{
\item{model}{a model object; lmerMod or lmerModLmerTest}
}
\value{
a list of contrast matrices for each model term.
}
\description{
Experimental - not extensively tested.
}
\author{
Rune Haubo B. Christensen
}
\keyword{internal}
