% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_contrasts.R
\name{get_contrasts_type3old}
\alias{get_contrasts_type3old}
\title{Compute Type III Contrast Matrices}
\usage{
get_contrasts_type3old(model)
}
\arguments{
\item{model}{a model object; lmerMod or lmerModLmerTest}
}
\value{
a list of contrast matrices for each model term.
}
\description{
Experimental - not extensively tested.
}
\author{
Rune Haubo B. Christensen
}
\keyword{internal}
