% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer.R
\name{get_covbeta}
\alias{get_covbeta}
\title{Compute cov(beta) as a Function of varpar of an LMM}
\usage{
get_covbeta(varpar, devfun)
}
\arguments{
\item{varpar}{variance parameters; \code{varpar = c(theta, sigma)}.}

\item{devfun}{deviance function as a function of theta only.}
}
\value{
cov(beta) at supplied varpar values.
}
\description{
At the optimum cov(beta) is available as vcov(lmer-model). This function
computes cov(beta) at non (RE)ML estimates of \code{varpar}.
}
\author{
Rune Haubo B. Christensen
}
\keyword{internal}
