% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contest.R
\name{get_Fstat_ddf}
\alias{get_Fstat_ddf}
\title{Compute denominator df for F-test}
\usage{
get_Fstat_ddf(nu, tol = 1e-08)
}
\arguments{
\item{nu}{vector of denominator df for the t-statistics}

\item{tol}{tolerance on the consequtive differences between elements of nu to}
}
\value{
the denominator df; a numerical scalar
}
\description{
From a vector of denominator df from independent t-statistics (\code{nu}),
the denominator df for the corresponding F-test is computed.
}
\details{
Note that if any \code{nu <= 2} then \code{2} is returned. Also, if all nu
are within tol of each other the simple average of the nu-vector is returned.
This is to avoid downward bias.
}
\author{
Rune Haubo B. Christensen
}
\keyword{internal}
