% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{get_model}
\alias{get_model}
\title{Extract Model from an Object}
\usage{
get_model(x, ...)
}
\arguments{
\item{x}{an object.}

\item{...}{currently not used.}
}
\description{
Extract Model from an Object
}
\seealso{
\code{\link{get_model.step_list}}
}
\keyword{internal}
