% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_means.R
\name{ls_means}
\alias{ls_means}
\alias{difflsmeans}
\alias{lsmeansLT}
\title{LS-means Generic Function}
\usage{
ls_means(model, ...)

difflsmeans(model, ...)

lsmeansLT(model, ...)
}
\arguments{
\item{model}{a model object.}

\item{...}{parsed on to methods.}
}
\description{
LS-means Generic Function
}
\seealso{
\code{\link{ls_means.lmerModLmerTest}}

\code{\link{difflsmeans.lmerModLmerTest}}

\code{\link{lsmeansLT.lmerModLmerTest}}
}
\author{
Rune Haubo B. Christensen
}
\keyword{internal}
