% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsmeans.R
\name{lsmeansLT}
\alias{lsmeansLT}
\title{LS-means for lmerTest Model Fits}
\usage{
lsmeansLT(model, which = NULL, level = 0.95, ddf = c("Satterthwaite",
  "KR"))
}
\arguments{
\item{model}{a model object fitted with \code{\link{lmer}} (of class
\code{"lmerModLmerTest"}).}

\item{which}{optional character vector naming factors for which LS-means should
be computed. If \code{NULL} (default) LS-means for all factors are computed.}

\item{level}{confidence level.}

\item{ddf}{method for computation of denominator degrees of freedom.}
}
\value{
An LS-means table in the form of a \code{data.frame}. Formally an object
of class \code{c("lsmeansLT", "data.frame")} with a number of attributes set.
}
\description{
Computes LS-means for all factors in a linear mixed model.
}
\details{
Confidence intervals and p-values are based on the t-distribution using
degrees of freedom based on Satterthwaites or Kenward-Roger methods.

LS-means is SAS terminology for predicted/estimated marginal means, i.e. means
for levels of factors which are averaged over the levels of other factors in
the model. A flat (i.e. unweighted) average is taken which gives equal weight
to all levels of each of the other factors. Numeric/continuous variables are
set at their mean values. See \pkg{emmeans} package
for more options and greater flexibility.

LS-means contrasts are checked for estimability and unestimable contrasts appear
as \code{NA}s in the resulting table.

LS-means objects (of class \code{"lsmeansLT"} have a print method).
}
\examples{

data("cake", package="lme4")
model <- lmer(angle ~ recipe * temp + (1|recipe:replicate), cake)
lsmeansLT(model)

}
\author{
Rune Haubo B. Christensen
}
