% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\docType{class}
\name{merModLmerTest-class}
\alias{merModLmerTest-class}
\alias{merModLmerTest}
\title{Legacy lmerTest representation of Linear Mixed-Effects Models}
\description{
The \code{merModLmerTest} class extends \code{lmerMod} (which extends
\code{merMod}) from the \pkg{lme4}-package.
}
\author{
Rune Haubo B. Christensen
}
\keyword{internal}
