% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{qform}
\alias{qform}
\title{Compute Quadratic Form}
\usage{
qform(x, A)
}
\arguments{
\item{x}{a numeric vector}

\item{A}{a symmetric numeric matrix}
}
\value{
a numerical scalar
}
\description{
Efficiently computes \eqn{x' A x} - or in R-notation:
}
\details{
Length of \code{x} should equal the number of rows and columns of \code{A}.
}
\keyword{internal}
