% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast_utils.R
\name{relatives}
\alias{relatives}
\title{Find which model terms 'Contain' a term}
\usage{
relatives(classes.term, term, term_names, factors)
}
\arguments{
\item{classes.term}{?}

\item{term}{?}

\item{term_names}{?}

\item{factors}{?}
}
\value{
?
}
\description{
Experimental - not extensively tested in this context.
}
\author{
Lifted from \code{lmerTest:::makeContrastType3SAS}
}
\keyword{internal}
