% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranova.R
\name{rm_complete_terms}
\alias{rm_complete_terms}
\title{Remove Terms from Formula}
\usage{
rm_complete_terms(terms, full_formula, random = TRUE)
}
\arguments{
\item{terms}{character vector (or list) of terms to remove from
\code{full_formula}}

\item{full_formula}{formula}

\item{random}{if \code{TRUE} names of the return list have parentheses around
them.}
}
\description{
Remove fixef or ranef terms from formula, return a list of modified formulae
with environment restored to that of the original formula.
}
\keyword{internal}
