% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer_anova.R
\name{single_anova}
\alias{single_anova}
\title{ANOVA Tables for Linear Mixed Models}
\usage{
single_anova(object, type = c("I", "II", "III", "1", "2", "2b", "3", "3b",
  "3c", "marginal"), ddf = c("Satterthwaite", "KR"))
}
\arguments{
\item{object}{an \code{lmerModLmerTest} object; the result of \code{lmer()}
after loading the \pkg{lmerTestR}-package.}

\item{type}{the type of ANOVA table requested (using the SAS terminology for
these) with Type I being the familiar sequential ANOVA table.}

\item{ddf}{method for computing denominator degrees of freedom.}
}
\value{
an ANOVA table
}
\description{
ANOVA Tables for Linear Mixed Models
}
\author{
Rune Haubo B. Christensen
}
\keyword{internal}
