% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step.R
\name{step}
\alias{step}
\alias{step.default}
\alias{step.lmerModLmerTest}
\alias{get_model}
\alias{get_model.step_list}
\title{Backward Elimination for Linear Mixed Models}
\usage{
step(object, ...)

\method{step}{default}(object, ...)

\method{step}{lmerModLmerTest}(object, ddf = c("Satterthwaite", "KR"),
  alpha.random = 0.1, alpha.fixed = 0.05, ...)

get_model(x, ...)

\method{get_model}{step_list}(x, ...)
}
\arguments{
\item{object}{the model}

\item{...}{currently not used}

\item{ddf}{degrees of freedom}

\item{alpha.random}{alpha for random effects}

\item{alpha.fixed}{alpha for fixed effects}

\item{x}{a step object}
}
\value{
a list ...
}
\description{
Backward Elimination for Linear Mixed Models
}
