% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contest.R
\name{contest}
\alias{contest}
\alias{contest1D}
\alias{contestMD}
\title{Generic Contrast Test Functions}
\usage{
contest(model, L, ...)

contest1D(model, L, ...)

contestMD(model, L, ...)
}
\arguments{
\item{model}{a model object.}

\item{L}{a contrast vector or matrix.}

\item{...}{additional arguments passed to methods.}
}
\description{
Generic functions for tests contrasts.
}
\seealso{
contest methods for \code{\link{lmer}} objects:
\code{\link[=contest.lmerModLmerTest]{contest}},
\code{\link[=contest1D.lmerModLmerTest]{contest1D}}, and
\code{\link[=contestMD.lmerModLmerTest]{contestMD}}.
}
\author{
Rune Haubo B. Christensen
}
\keyword{internal}
