% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_contrasts.R
\name{get_contrasts_type1}
\alias{get_contrasts_type1}
\title{Type I ANOVA table contrasts}
\usage{
get_contrasts_type1(model)
}
\arguments{
\item{model}{a model object with \code{terms} and \code{model.matrix} methods.}
}
\value{
List of contrast matrices - one contrast matrix for each model term.
}
\description{
Type I ANOVA table contrasts
}
\author{
Rune Haubo B. Christensen
}
\keyword{internal}
