% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_contrasts.R
\name{get_contrasts_type3}
\alias{get_contrasts_type3}
\title{Contrasts for Type III Tests}
\usage{
get_contrasts_type3(model, which = NULL)
}
\arguments{
\item{model}{model object.}

\item{which}{optional character vector naming terms for which to compute the
the contrasts.}
}
\value{
list of contrast matrices.
}
\description{
Contrasts for Type III Tests
}
\keyword{internal}
