% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type3.R
\name{get_rdX}
\alias{get_rdX}
\title{Compute the 'Full' Rank-Deficient Design Matrix}
\usage{
get_rdX(model, do.warn = TRUE, use_contrasts = FALSE)
}
\arguments{
\item{model}{a model object; lmerMod or lmerModLmerTest.}

\item{do.warn}{throw a warning if there is no data for some factor
combinations.}

\item{use_contrasts}{if \code{TRUE} contrasts coding from the original design
matrix will be used. If \code{FALSE} the settings in \code{options()$contrasts}
will be used.}
}
\value{
the rank-deficien design matrix, \code{rdX} with attributes
\item{param}{a vector of _parameters_ in the over-parameterized model
notation with length equal to the number of columns in \code{rdX}.}
\item{is_coef}{numeric vector with indices on the columns in \code{rdX}
which corresponds to the columns in the original full-rank \code{X} and
therefore also to the coefficient vector.}
}
\description{
Experimental - not extensively tested.
}
\author{
Rune Haubo B. Christensen
}
\keyword{internal}
