% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast_utils.R
\name{get_rdX}
\alias{get_rdX}
\title{Compute the 'Full' Rank-Deficient Design Matrix}
\usage{
get_rdX(model, do.warn = TRUE)
}
\arguments{
\item{model}{a model object; lmerMod or lmerModLmerTest.}

\item{do.warn}{throw a message if there is no data for some factor
combinations.}
}
\value{
the rank-deficien design matrix
}
\description{
Compute the 'Full' Rank-Deficient Design Matrix
}
\author{
Rune Haubo B. Christensen
}
\keyword{internal}
