% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsmeans.R
\name{show_contrasts}
\alias{show_contrasts}
\title{Show LS-means Contrasts}
\usage{
show_contrasts(object, fractions = FALSE, names = TRUE)
}
\arguments{
\item{object}{an \code{lsmeans} object.}

\item{fractions}{display contrasts as fractions rather than decimal numbers?}

\item{names}{include row and column names of the contrasts matrices?}
}
\value{
a list of contrast matrices; one matrix for each model term.
}
\description{
Extracts the contrasts which defines the LS-mean contrasts.
}
\examples{

data("cake", package="lme4")
model <- lmer(angle ~ recipe * temp + (1|recipe:replicate), cake)
(lsm <- lsmeans(model))
show_contrasts(lsm)

}
\seealso{
\code{\link{lsmeans}} for computation of LS-means.
}
