% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer_anova.R
\name{show_tests}
\alias{show_tests}
\alias{show_tests.default}
\title{Show Tests Generic Function and Default Method}
\usage{
show_tests(object, ...)

\method{show_tests}{default}(object, fractions = FALSE, names = TRUE, ...)
}
\arguments{
\item{object}{a suitable object with an \code{"hypotheses"} attribute, e.g. an
anova table or an \code{ls_means} table as defined in \pkg{lmerTest}.}

\item{...}{parsed on to methods; currently not used in the default method.}

\item{fractions}{display entries in the hypothesis matrices as fractions?}

\item{names}{if \code{FALSE} column and row names of the hypothesis matrices
are suppressed.}
}
\description{
Show Tests Generic Function and Default Method
}
\seealso{
\code{\link{show_tests.anova}} and \code{\link{show_tests.ls_means}}
}
\author{
Rune Haubo B. Christensen

Rune Haubo B. Christensen
}
\keyword{internal}
