% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmer_anova.R
\name{single_anova}
\alias{single_anova}
\title{ANOVA Tables for Linear Mixed Models}
\usage{
single_anova(
  object,
  type = c("III", "II", "I", "3", "2", "1", "yates", "marginal", "2b"),
  ddf = c("Satterthwaite", "Kenward-Roger")
)
}
\arguments{
\item{object}{an \code{lmerModLmerTest} object; the result of \code{lmer()}
after loading the \pkg{lmerTest}-package.}

\item{type}{the type of ANOVA table requested (using the SAS terminology for
these) with Type I being the familiar sequential ANOVA table.}

\item{ddf}{method for computing denominator degrees of freedom.}
}
\value{
an ANOVA table
}
\description{
ANOVA Tables for Linear Mixed Models
}
\author{
Rune Haubo B. Christensen
}
\keyword{internal}
